New config of Paradiseo with Cmake 2.6
git-svn-id: svn://scm.gforge.inria.fr/svnroot/paradiseo@1347 331e1502-861f-0410-8da2-ba01fb791d7f

+++ trunk/paradiseo-moeo/src/fitness/moeoDominanceDepthFitnessAssignment.h
@@ -173,7 +173,7 @@ class moeoDominanceDepthFitnessAssignment : public moeoDominanceBasedFitnessAssi
       // assign fitness values
       unsigned int rank = 1;
       _pop[_pop.size()-1].fitness(rank);
-      for (int i=((int)_pop.size())-2; i>=0; i--)
+      for (int i=_pop.size()-2; i>=0; i--)
         {
           if (_pop[i].objectiveVector() != _pop[i+1].objectiveVector())
             {
